fix(embodiment): keep fp32 master weights for openpi_pytorch SFT training
The openpi_pytorch get_model factory cast the fp32-loaded training weights to
bf16 before FSDP wrapped them, so AdamW updated bf16 master weights and the tiny
warmup-LR updates (~1e-6, below the bf16 ULP near 1.0) were lost to rounding --
the first-50-step BEHAVIOR SFT loss stayed flat at ~0.24 while the reference
dropped 0.246->0.090. Keep the fp32 master for load_for_training and let FSDP
MixedPrecision cast to bf16 only for compute (matching the reference recipe:
fp32 load + FSDP1 MixedPrecision); eval stays bf16-strict.

After the fix the first-50 loss descends 0.226->0.046 (46/50 within the DEC-1(b)
band, 29/50 within |d|<=0.03; was 12/50). A controlled rank-independent re-run is
near-identical to production, confirming AC-6 rank-folding is not the residual.

- get_model: skip the bf16 cast for training; refresh the docstring
- test_checkpoint: training model is now fp32 master (was bf16)
- test_sft_loss_parity_gpu: cast to bf16 for the non-FSDP forward-parity check
- docs: first-50 evidence + r20 scalar CSVs (production + rank-independent control)

+++ tests/unit_tests/test_openpi_pytorch_sft_loss_parity_gpu.py
@@ -121,7 +121,11 @@ def _rlinf_per_batch_losses(out):
             },
         }
     )
-    model = get_model(model_cfg).to("cuda")
+    # The training factory keeps fp32 master weights; FSDP MixedPrecision casts to
+    # bf16 for compute in production. There is no FSDP wrapper here, so cast the
+    # model to bf16 to reproduce the production bf16 forward (and the model's
+    # bf16 embed_dtype) for this same-batch forward-parity check.
+    model = get_model(model_cfg).to("cuda").to(torch.bfloat16)
     model.train()
     inner = model.model
